pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"


[project]
name = "nuvolos-cli"
dynamic = ["version"]
description = "Command-line interface for Nuvolos"
authors = [
    {name = "Alphacruncher", email = "support@nuvolos.cloud"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "click>=8.1.7",
    "click-log>=0.4.0",
    "tabulate>=0.9.0",
    "pyyaml>=6.0.1",
    "nuvolos-client-api=1.2.1",
    "humanize>=4.9.0",
    "python-slugify>=8.0.4",
]

[tool.setuptools_scm]


[project.urls]
Repository = "https://github.com/nuvolos-cloud/nuvolos-cli"
Documentation = "https://nuvolos-cli.readthedocs.io/en/latest"

[project.scripts]
nuvolos = "nuvolos_cli.interface:nuvolos"